Rename namespaces and organize imports
