+++ src/Argon.Api/Controllers/ChannelsController.cs
@@ -1,11 +1,11 @@
 namespace Argon.Api.Controllers;
 
 using Attributes;
-using Entities;
-using Grains.Interfaces;
+using DataTypes;
+using global::Grains.Interface;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Sfu;
+using Models;
 using Swashbuckle.AspNetCore.Annotations;
 
 #if DEBUG
