+++ src/Argon.Api/Controllers/ServersController.cs
@@ -1,10 +1,10 @@
 namespace Argon.Api.Controllers;
 #if DEBUG
 using Attributes;
-using Entities;
-using Grains.Interfaces;
+using global::Grains.Interface;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Models;
 using Swashbuckle.AspNetCore.Annotations;
 
 [Route("api/[controller]")]
