+++ src/Argon.Api/Entities/ApplicationDbContext.cs
@@ -1,6 +1,7 @@
 namespace Argon.Api.Entities;
 
 using Microsoft.EntityFrameworkCore;
+using Models;
 
 public class ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : DbContext(options)
 {
