+++ src/Argon.Api/Features/Sfu/IArgonSelectiveForwardingUnit.cs
@@ -1,4 +1,6 @@
-namespace Argon.Sfu;
+namespace Argon.Api.Features.Sfu;
+
+using DataTypes;
 
 public interface IArgonSelectiveForwardingUnit
 {
