+++ src/Argon.Api/Grains/ChannelManager.cs
@@ -1,10 +1,12 @@
 namespace Argon.Api.Grains;
 
+using DataTypes;
 using Entities;
-using Interfaces;
+using Features.Sfu;
+using global::Grains.Interface;
 using Microsoft.EntityFrameworkCore;
+using Models;
 using Persistence.States;
-using Sfu;
 
 public class ChannelManager(
     IArgonSelectiveForwardingUnit sfu,
