+++ src/Argon.Api/Program.cs
@@ -7,10 +7,11 @@ using Argon.Api.Features.EmailForms;
 using Argon.Api.Features.Jwt;
 using Argon.Api.Features.Otp;
 using Argon.Api.Features.Rpc;
-using Argon.Api.Grains.Interfaces;
+using Argon.Api.Features.Sfu;
 using Argon.Api.Migrations;
 using Argon.Api.Services;
-using Argon.Sfu;
+using Grains.Interface;
+using ServiceDefaults;
 
 var builder = WebApplication.CreateBuilder(args);
